requirements.txt
aiogram
asyncio
schedule
python-dotenv
apscheduler
pytz
asyncpg